added toggle to show network IDs, fixed assignment bug
if the last neighbor of a node transitioned to initialized, the returned nextId would match the id sent to it. This would result in tableChange being set to false, making the parent transition to initialized, even if a previous child did have a change. Now  tableChange value is OR-ed with nextId comparison